==> goal_kick_comparative_analysis/progress_files/04_split+loop_possession_chains.ipynb <==
import pandas as pd
import seaborn as sns
import numpy as np

# accessing data library
from statsbombpy import sb

# opening the libraries as DF's
from mplsoccer import Sbopen
parser = Sbopen()

# drawing pitches 
import matplotlib.pyplot as plt
from mplsoccer import Pitch, VerticalPitch

## data prep ideas for avg. calculations

+ concat all matches from all seasons
    + competitions.head(60)
    + WSL = competition_id: 37
    + seasons = season_id: 90 (2020/21), 42 (2019/20), 4 (2018/19)

# preparing data to be iterated... 

## LOAD IN MATCH DATA ##
df = sb.events(match_id=3764234)

## Create out-of-possession-team column, called 'opposing_team' ##
home_team = df['team'].iloc[0]
away_team = df['team'].iloc[1]

def out_of_possession_team(row):
    if row['possession_team'] == home_team:
        return away_team
    elif row['possession_team'] == away_team:
        return home_team
    else:
        return None

df['opposing_team'] = df.apply(out_of_possession_team, axis=1)


## Split the location info into X,Y (and Z for shot) columns ##
df[['X_start', 'Y_start']] = df['location'].apply(pd.Series)
df[['X_end_pass', 'Y_end_pass']] = df['pass_end_location'].apply(pd.Series)
df[['X_end_carry', 'Y_end_carry']] = df['carry_end_location'].apply(pd.Series)
df[['X_end_shot', 'Y_end_shot', 'Z_end_shot']] = df['shot_end_location'].apply(pd.Series)
df[['X_end_gk', 'Y_end_gk']] = df['goalkeeper_end_location'].apply(pd.Series)


## Narrow the DF to only neccessary columns ##
df = df[['timestamp', 'duration','possession','possession_team','opposing_team',
         'play_pattern','type','team','player','position',
         'pass_length','pass_angle','pass_height','pass_outcome','pass_body_part','pass_shot_assist',
         'shot_statsbomb_xg','shot_outcome',
         'X_start', 'Y_start',
         'X_end_pass','Y_end_pass','X_end_carry','Y_end_carry','X_end_shot','Y_end_shot','Z_end_shot']]

## Narrow the DF to only neccessary rows (based on 'type') ##
df = df.loc[(chain_16["type"] == "Pass") | 
            (chain_16["type"] == "Carry") | 
            (chain_16["type"] == 'Shot')]

gk_chains = df[(df['play_pattern']=='From Goal Kick')]['possession'].unique()

## checking data before iteration

df = sb.events(match_id=3764234)
df_match = df.copy()

## obtain the home / away teams from the data ##

df_match[df_match['type'] == 'Starting XI'][['team']]

home_team = df_match['team'].iloc[0]
away_team = df_match['team'].iloc[1]

type(home_team)

df_match['possession_team'].describe(include = 'O')

df_match[df_match['possession_team'] == home_team][['possession_team']]

## create new column detailing the out-of-possesion team

def out_of_possession_team(row):
    if row['possession_team'] == home_team:
        return away_team
    elif row['possession_team'] == away_team:
        return home_team
    else:
        return None

df_match['opposing_team'] = df_match.apply(out_of_possession_team, axis=1)


df_match[['possession_team','opposing_team']]

df_match.columns

df_match['shot_outcome'].unique()

df_match['play_pattern'].unique()

# setting up the x_start & y_start columns

df_match[['X_start', 'Y_start']] = df_match['location'].apply(pd.Series)
df_match.columns

# setting up the x_end, y_end columns

df_match[['X_end_pass', 'Y_end_pass']] = df_match['pass_end_location'].apply(pd.Series)
df_match[['X_end_carry', 'Y_end_carry']] = df_match['carry_end_location'].apply(pd.Series)
df_match[['X_end_shot', 'Y_end_shot', 'Z_end_shot']] = df_match['shot_end_location'].apply(pd.Series)
df_match[['X_end_gk', 'Y_end_gk']] = df_match['goalkeeper_end_location'].apply(pd.Series)

df_match.columns

type(df_match['pass_end_location'])

# Combine 'x_end' columns into one column 'x_end'
X_end_columns = [col for col in df_match.columns if 'X_end' in col]
df_match['x_end'] = df_match[X_end_columns].apply(lambda row: row.dropna().iloc[0] if not row.dropna().empty else np.nan, axis=1)


# Combine 'y_end' columns into one column 'y_end'
Y_end_columns = [col for col in df_match.columns if 'Y_end' in col]
df_match['y_end'] = df_match[Y_end_columns].apply(lambda row: row.dropna().iloc[0] if not row.dropna().empty else np.nan, axis=1)

# Drop the original 'x_end' and 'y_end' columns
df_match = df_match.drop(columns=X_end_columns + Y_end_columns)
df_match.columns

df_match[['x_end']].describe()

home_team = df_match['team'].iloc[0]
away_team = df_match['team'].iloc[1]

def out_of_possession_team(row):
    if row['possession_team'] == home_team:
        return away_team
    elif row['possession_team'] == away_team:
        return home_team
    else:
        return None

df_match['defending_team'] = df_match.apply(out_of_possession_team, axis=1)

## create split DF's for each team ##
df_home_team = df[df['possession_team'] == home_team]
df_away_team = df[df['possession_team'] == away_team]

type(df_match[['pass_end_location']].iloc[36])
