==> goal_kick_comparative_analysis/progress_files/05_working_notes.ipynb <==
import pandas as pd
import numpy as np
import json
import seaborn as sns

# accessing data library
from statsbombpy import sb

# opening the libraries as DF's
from mplsoccer import Sbopen
parser = Sbopen()

# plotting
import os
import pathlib
import warnings
import statsmodels.api as sm
import statsmodels.formula.api as smf
from mplsoccer import Pitch, VerticalPitch
import matplotlib.pyplot as plt

pd.options.mode.chained_assignment = None
warnings.filterwarnings('ignore')

# What is a Possession Chain?


| Variable | Variable Type | Variable Description | Values | Value Description |
|----------|---------------|----------------------|--------|-------------------|
| possession | integer | Indicates the current unique possession in the game. A single possession denotes a period of play in which the ball is in play and a single team is in control of the ball. | e.g., 1 - # of unique possessions | New possession are triggered after a team demonstrate they've established control of the ball. A new possession can begin even if the same team has possession of the ball for example, a blocked pass goes out for a throw in for the same team, this would be a new possession for the same attacking team.

# Working with the WSL, the competition_id = 37

## The relevant season_id will be 90 (20/21), 42 (19/20), and 4 (18/19)


competitions = sb.competitions()

competitions[competitions['competition_name']=="FA Women's Super League"]

# There are 130 games for each season
## I'll need to iterate over each game in each season, and add a column to describe each for the event data.

# Below are the list of match_id 's for each of the seasons

wsl_18_19 = sb.matches(competition_id=37,season_id=4)
match_list_18_19 = wsl_18_19.sort_values(['match_date','kick_off','match_id'])['match_id'].tolist()


wsl_19_20 = sb.matches(competition_id=37,season_id=42)
match_list_19_20 = wsl_19_20.sort_values(['match_date','kick_off','match_id'])['match_id'].tolist()

wsl_20_21 = sb.matches(competition_id=37,season_id=90)
match_list_20_21 = wsl_20_21.sort_values(['match_date','kick_off','match_id'])['match_id'].tolist()

#match_list_18_19
#match_list_19_20
#match_list_20_21


# Below is how I access specific matches

match_events = sb.events(match_id=3764234)

match_events.shape

match_events.columns

match_events[match_events[
    'possession']==16].loc[match_events[
    'type']=='Pass'][['timestamp','possession_team','team','player','position',
                      'pass_recipient', 'pass_outcome','pass_miscommunication',
                      'location','pass_end_location']]

+ # I want to be able to concatenate all games across a season to then analyse each team's performance for the whole season.

+ # Firstly, I only want to consider the possesion chains beginning with a GK as well as the chain following:

```python
## ONLY GK CHAINS ##
gk_chains = (df[(df['play_pattern']=='From Goal Kick')]['possession'].unique())
gk_chain_list = gk_chains.tolist()

## THE CHAIN FOLLOWING A GK CHAIN ##
aftr_gk_chain_list = (gk_chains+1).tolist()
aftr_gk_chain_list
````

+ ## Once I have event data for a single game, I want to filter to the relevant chains, add columns for 'season', 'match', 'home_team', 'away_team', and 'defending_team' team:

```python

## CREATING THE DEFENDING TEAM COLUMN ##

home_team = df_match['team'].iloc[0]
away_team = df_match['team'].iloc[1]

def out_of_possession_team(row):
    if row['possession_team'] == home_team:
        return away_team
    elif row['possession_team'] == away_team:
        return home_team
    else:
        return None

df_match['defending_team'] = df_match.apply(out_of_possession_team, axis=1)

```

+ ## I then want to move onto the next game's event data and repeat until end.

+ ## I'll also need to split the start & end locations for each event (or 'type):
```python
df[['X_start', 'Y_start']] = df['location'].apply(pd.Series)
df[['X_end_pass', 'Y_end_pass']] = df['pass_end_location'].apply(pd.Series)
df[['X_end_carry', 'Y_end_carry']] = df['carry_end_location'].apply(pd.Series)
df[['X_end_shot', 'Y_end_shot', 'Z_end_shot']] = df['shot_end_location'].apply(pd.Series)
df[['X_end_gk', 'Y_end_gk']] = df['goalkeeper_end_location'].apply(pd.Series)
````
Is it possible to fit the end locations into single x, y, z columns, oppose to one for each event type?

+ ## Once completed, I want to narrow the columns to only the relevant...
... Currently, this would be:
        

```python

df[['match','season','home_team','away_team',
    'timestamp', 'duration','possession','possession_team','defending_team',
    'play_pattern','type','team','player','position',
    'pass_length','pass_angle','pass_height','pass_outcome','pass_body_part','pass_shot_assist',
    'shot_statsbomb_xg','shot_outcome',
    'X_start', 'Y_start',
    'X_end_pass','Y_end_pass','X_end_carry','Y_end_carry','X_end_shot','Y_end_shot','Z_end_shot']]
         
````


match_events['play_pattern'].unique()

## VIEWING THE STARTING 'TYPE' OF EVERY POSSESSION CHAIN TO UNDERSTAND WHICH TYPES ARE NEEDED...

#match_events.sort_values(by='timestamp').groupby('possession').first()['type'].unique()

#match_events.groupby('possession').last()['type'].unique()
match_events.groupby('possession').first()['type'].unique()


## VIEWING THE STARTING 'PLAY_PATTERN' OF EVERY POSSESSION CHAIN...

#match_events.sort_values(by='timestamp').groupby('possession').first()['play_pattern'].unique()
#match_events.groupby('possession').first()['play_pattern'].unique()

# Creating a means of concatenating the match event data


## Creating the match lists for each season:

wsl_18_19 = sb.matches(competition_id=37,season_id=4)
match_list_18_19 = wsl_18_19.sort_values(['match_date','kick_off','match_id'])['match_id'].tolist()


wsl_19_20 = sb.matches(competition_id=37,season_id=42)
match_list_19_20 = wsl_19_20.sort_values(['match_date','kick_off','match_id'])['match_id'].tolist()

wsl_20_21 = sb.matches(competition_id=37,season_id=90)
match_list_20_21 = wsl_20_21.sort_values(['match_date','kick_off','match_id'])['match_id'].tolist()

#match_list_18_19
#match_list_19_20
#match_list_20_21

# match_events = sb.events(match_id=3764234

# Creating the function to output match event data:

match = 19770
season = '18/19'

def parse_match_data(match): ## SETTING-UP HOME & AWAY TEAMS AS VARIABLES
    df_match = sb.events(match_id = match)
    home_team = df_match['team'].iloc[0]
    away_team = df_match['team'].iloc[1]

    def out_of_possession_team(row): ## CREATING THE DEFENDING TEAM COLUMN
        if row['possession_team'] == home_team:
            return away_team
        elif row['possession_team'] == away_team:
            return home_team
        else:
            return None

    df_match['defending_team'] = df_match.apply(out_of_possession_team, axis=1)
    
    ## SPLITTING THE START LOACTION INTO X,Y COLUMNS
    df_match[['x_start', 'y_start']] = df_match['location'].apply(pd.Series)
    
    ## SPLITTING THE END LOACTIONS INTO X,Y COLUMNS
    df_match[['X_end_pass', 'Y_end_pass']] = df_match['pass_end_location'].apply(pd.Series)
    df_match[['X_end_carry', 'Y_end_carry']] = df_match['carry_end_location'].apply(pd.Series)
    df_match[['X_end_shot', 'Y_end_shot', 'z_end_shot']] = df_match['shot_end_location'].apply(pd.Series)
    df_match[['X_end_gk', 'Y_end_gk']] = df_match['goalkeeper_end_location'].apply(pd.Series)
    
    ## Combine 'x_end' columns into one column 'x_end'
    X_end_columns = [col for col in df_match.columns if 'X_end' in col]
    df_match['x_end'] = df_match[X_end_columns].apply(lambda row: row.dropna().iloc[0] if not row.dropna().empty else np.nan, axis=1)

    ## Combine 'y_end' columns into one column 'y_end'
    Y_end_columns = [col for col in df_match.columns if 'Y_end' in col]
    df_match['y_end'] = df_match[Y_end_columns].apply(lambda row: row.dropna().iloc[0] if not row.dropna().empty else np.nan, axis=1)
    
    ## Drop the original 'x_end' and 'y_end' columns
    df_match = df_match.drop(columns=X_end_columns + Y_end_columns)
    
    df_match['home_team'] = home_team
    df_match['away_team'] = away_team
    
    df_match['match'] = match
    df_match['season'] = '2018/19'
    
    df_match = df_match[['match','season','home_team','away_team',
    'timestamp', 'duration','possession','possession_team','defending_team',
    'play_pattern','type','team','player','position',
    'pass_recipient','pass_outcome','pass_length','pass_height','pass_angle','pass_body_part','pass_shot_assist',
    'shot_statsbomb_xg','shot_outcome',
    'x_start', 'y_start','x_end','y_end','z_end_shot']]
        
    return df_match

# Data Check...

parse_match_data(19777).columns

## Check actual data...

sb.events(match_id = 19777)[['type']]

## Check parsed data...

parse_match_data(19777)[['type']]

parse_match_data(19777).groupby('possession').first()['type'].unique()

parse_match_data(19777)[parse_match_data(19777)['type']=='Referee Ball-Drop']

parse_match_data(19777)[parse_match_data(19777)['possession']==169].sort_values(by='timestamp')

# Concatenating a season's matches

wsl_18_19 = sb.matches(competition_id=37,season_id=4)
match_list_18_19 = wsl_18_19.sort_values(['match_date','kick_off','match_id'])['match_id'].tolist()


wsl_19_20 = sb.matches(competition_id=37,season_id=42)
match_list_19_20 = wsl_19_20.sort_values(['match_date','kick_off','match_id'])['match_id'].tolist()

wsl_20_21 = sb.matches(competition_id=37,season_id=90)
match_list_20_21 = wsl_20_21.sort_values(['match_date','kick_off','match_id'])['match_id'].tolist()

#match_list_18_19
#match_list_19_20
#match_list_20_21

def parse_match_data(match): ## SETTING-UP HOME & AWAY TEAMS AS VARIABLES
    df_match = sb.events(match_id = match)
    home_team = df_match['team'].iloc[0]
    away_team = df_match['team'].iloc[1]

    def out_of_possession_team(row): ## CREATING THE DEFENDING TEAM COLUMN
        if row['possession_team'] == home_team:
            return away_team
        elif row['possession_team'] == away_team:
            return home_team
        else:
            return None

    df_match['defending_team'] = df_match.apply(out_of_possession_team, axis=1)
    
    ## SPLITTING THE START LOACTION INTO X,Y COLUMNS
    df_match[['x_start', 'y_start']] = df_match['location'].apply(pd.Series)
    
    ## SPLITTING THE END LOACTIONS INTO X,Y COLUMNS
    df_match[['X_end_pass', 'Y_end_pass']] = df_match['pass_end_location'].apply(pd.Series)
    df_match[['X_end_carry', 'Y_end_carry']] = df_match['carry_end_location'].apply(pd.Series)
    df_match[['X_end_shot', 'Y_end_shot', 'z_end_shot']] = df_match['shot_end_location'].apply(pd.Series)
    df_match[['X_end_gk', 'Y_end_gk']] = df_match['goalkeeper_end_location'].apply(pd.Series)
    
    ## Combine 'x_end' columns into one column 'x_end'
    X_end_columns = [col for col in df_match.columns if 'X_end' in col]
    df_match['x_end'] = df_match[X_end_columns].apply(lambda row: row.dropna().iloc[0] if not row.dropna().empty else np.nan, axis=1)

    ## Combine 'y_end' columns into one column 'y_end'
    Y_end_columns = [col for col in df_match.columns if 'Y_end' in col]
    df_match['y_end'] = df_match[Y_end_columns].apply(lambda row: row.dropna().iloc[0] if not row.dropna().empty else np.nan, axis=1)
    
    ## Drop the original 'x_end' and 'y_end' columns
    df_match = df_match.drop(columns=X_end_columns + Y_end_columns)
    
    df_match['home_team'] = home_team
    df_match['away_team'] = away_team
    
    df_match['match'] = match
    
    df_match = df_match[['match','home_team','away_team',
    'timestamp', 'duration','possession','possession_team','defending_team',
    'play_pattern','type','team','player','position',
    'pass_recipient','pass_outcome','pass_length','pass_height','pass_angle','pass_body_part','pass_shot_assist',
    'shot_statsbomb_xg','shot_outcome',
    'x_start', 'y_start','x_end','y_end','z_end_shot']]
        
    return df_match

# 2018-19 DF

matches_ssn1819 = pd.DataFrame()

for match in match_list_18_19:
    # Call the parse_match_data function for each match_id
    match_data = parse_match_data(match)
    
    # Concatenate the current match_data to the overall dataframe
    matches_ssn1819 = pd.concat([matches_ssn1819, match_data], ignore_index=True)
    
    matches_ssn1819['season'] = '2018/19'

matches_ssn1819.shape

matches_ssn1819.columns

matches_ssn1819 = matches_ssn1819[['season','match', 'home_team', 'away_team', 'timestamp', 'duration',
       'possession', 'possession_team', 'defending_team', 'play_pattern',
       'type', 'team', 'player', 'position', 'pass_length', 'pass_angle',
       'pass_recipient','pass_outcome','pass_length','pass_height','pass_angle','pass_body_part','pass_shot_assist',
       'shot_statsbomb_xg', 'shot_outcome', 'x_start', 'y_start', 'x_end',
       'y_end', 'z_end_shot']]

matches_ssn1819

matches_ssn1819['shot_outcome'].describe()

# 2019-20 DF

matches_ssn1920 = pd.DataFrame()

for match in match_list_19_20:
    match_data = parse_match_data(match)
    matches_ssn1920 = pd.concat([matches_ssn1920, match_data], ignore_index=True)
    matches_ssn1920['season'] = '2019/20'

matches_ssn1920 = matches_ssn1920[['season','match', 'home_team', 'away_team', 'timestamp', 'duration',
       'possession', 'possession_team', 'defending_team', 'play_pattern',
       'type', 'team', 'player', 'position', 'pass_length', 'pass_angle',
       'pass_recipient','pass_outcome','pass_length','pass_height','pass_angle','pass_body_part','pass_shot_assist',
       'shot_statsbomb_xg', 'shot_outcome', 'x_start', 'y_start', 'x_end',
       'y_end', 'z_end_shot']]

matches_ssn1920.columns

# 2020-21 DF

matches_ssn2021 = pd.DataFrame()

for match in match_list_20_21:
    match_data = parse_match_data(match)
    matches_ssn2021 = pd.concat([matches_ssn2021, match_data], ignore_index=True)
    matches_ssn2021['season'] = '2020/21'

matches_ssn2021 = matches_ssn2021[['season','match', 'home_team', 'away_team', 'timestamp', 'duration',
       'possession', 'possession_team', 'defending_team', 'play_pattern',
       'type', 'team', 'player', 'position', 'pass_length', 'pass_angle',
       'pass_recipient','pass_outcome','pass_length','pass_height','pass_angle','pass_body_part','pass_shot_assist',
       'shot_statsbomb_xg', 'shot_outcome', 'x_start', 'y_start', 'x_end',
       'y_end', 'z_end_shot']]

matches_ssn2021.shape

matches_ssn2021.to_csv('../data/matches_ssn2021.csv', index=False)

matches_ssn1920.to_csv('../data/matches_ssn1920.csv', index=False)

matches_ssn1819.to_csv('../data/matches_ssn1819.csv', index=False)
